requirements.txt
# Core Data Science
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Machine Learning - Traditional
scikit-learn>=1.3.0
xgboost>=2.0.0
catboost>=1.2.0
lightgbm>=4.0.0

# Deep Learning
torch>=2.0.0
tensorflow>=2.13.0
keras>=2.13.0

# Signal Processing & Filtering
pykalman>=0.9.5
PyWavelets>=1.4.1

# Data Handling
h5py>=3.8.0
tables>=3.8.0
pyarrow>=12.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Experiment Tracking
mlflow>=2.5.0
wandb>=0.15.0

# Progress & Utils
tqdm>=4.65.0
joblib>=1.3.0
pyyaml>=6.0

# API & Data Download
requests>=2.31.0
aiohttp>=3.8.0
ccxt>=4.0.0  # Cryptocurrency exchange API

# Statistical Analysis
statsmodels>=0.14.0
pingouin>=0.5.3  # Statistical tests

# LOB Specific (if available)
# lobframe  # You may need to install from source

# Development & Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.3.0
flake8>=6.0.0
ipython>=8.12.0
jupyter>=1.0.0
jupyterlab>=4.0.0

# Performance
numba>=0.57.0
